[CSL-1361] Move X.Arbitrary to Arbitrary.X
