+++ core/cardano-sl-core.cabal
@@ -30,10 +30,14 @@ flag embed-config
   description: Embed the config even if Development mode is enabled
 
 library
-  exposed-modules:     Pos.Core
+  exposed-modules:
+                       Pos.Arbitrary.Core
+                       Pos.Arbitrary.Core.Unsafe
+                       Pos.Arbitrary.Crypto
+                       Pos.Arbitrary.Crypto.Unsafe
+
+                       Pos.Core
                        Pos.Core.Address
-                       Pos.Core.Arbitrary
-                       Pos.Core.Arbitrary.Unsafe
                        Pos.Core.Block
                        Pos.Core.BlockVersionData
                        Pos.Core.Class
@@ -54,8 +58,6 @@ library
                        Pos.Core.Genesis.Types
 
                        Pos.Crypto
-                       Pos.Crypto.Arbitrary
-                       Pos.Crypto.Arbitrary.Unsafe
                        Pos.Crypto.Encryption
                        Pos.Crypto.AsBinary
                        Pos.Crypto.HD
