+++ godtossing/cardano-sl-godtossing.cabal
@@ -13,7 +13,7 @@ cabal-version:       >=1.10
 
 library
   exposed-modules:
-    Pos.Ssc.GodTossing.Arbitrary
+    Pos.Arbitrary.Ssc.GodTossing
 
     Pos.Ssc.GodTossing.Core
     Pos.Ssc.GodTossing.Core.Types
