+++ infra/Pos/Arbitrary/Slotting.hs
@@ -1,13 +1,13 @@
 -- | Arbitrary instances for Pos.Slotting types (infra package)
 
-module Pos.Slotting.Arbitrary () where
+module Pos.Arbitrary.Slotting () where
 
 import           Test.QuickCheck                   (Arbitrary (..))
 import           Test.QuickCheck.Arbitrary.Generic (genericArbitrary, genericShrink)
 
+import           Pos.Arbitrary.Core                ()
 import           Pos.Slotting.Types                (EpochSlottingData (..),
                                                     SlottingData (..))
-import           Pos.Core.Arbitrary                ()
 
 instance Arbitrary EpochSlottingData where
     arbitrary = genericArbitrary
