+++ src/Pos/Arbitrary/Txp/Network.hs
@@ -2,16 +2,16 @@
 
 -- | 'Arbitrary' instances for 'Pos.Txp.Network' types defined in 'src'
 
-module Pos.Txp.Network.Arbitrary () where
+module Pos.Arbitrary.Txp.Network () where
 
 import           Universum
 
 import           Test.QuickCheck                   (Arbitrary (..))
 import           Test.QuickCheck.Arbitrary.Generic (genericArbitrary, genericShrink)
 
+import           Pos.Arbitrary.Txp                 ()
 import           Pos.Binary.Update                 ()
 import           Pos.Communication.Types.Relay     (DataMsg (..))
-import           Pos.Txp.Arbitrary                 ()
 import           Pos.Txp.Network.Types             (TxMsgContents (..))
 
 instance Arbitrary TxMsgContents where
