+++ src/Pos/Txp/Network.hs
@@ -5,6 +5,6 @@ module Pos.Txp.Network
        , module Pos.Txp.Network.Types
        ) where
 
-import Pos.Txp.Network.Arbitrary ()
-import Pos.Txp.Network.Listeners
-import Pos.Txp.Network.Types
+import           Pos.Arbitrary.Txp.Network ()
+import           Pos.Txp.Network.Listeners
+import           Pos.Txp.Network.Types
