+++ ssc/cardano-sl-ssc.cabal
@@ -13,7 +13,7 @@ cabal-version:       >=1.10
 
 library
   exposed-modules:
-    Pos.Ssc.Arbitrary
+    Pos.Arbitrary.Ssc
 
     Pos.Ssc.Class
     Pos.Ssc.Class.Types
