+++ test/Test/Pos/Genesis/Identity/BinarySpec.hs
@@ -4,14 +4,14 @@ module Test.Pos.Genesis.Identity.BinarySpec
        ( spec
        ) where
 
-import           Test.Hspec                   (Spec, describe)
+import           Test.Hspec         (Spec, describe)
 import           Universum
 
-import qualified Pos.Genesis                  as T
-import           Pos.Core.Arbitrary           ()
-import           Pos.Ssc.GodTossing           ()
+import           Pos.Arbitrary.Core ()
+import qualified Pos.Genesis        as T
+import           Pos.Ssc.GodTossing ()
 
-import           Test.Pos.Util                (binaryTest)
+import           Test.Pos.Util      (binaryTest)
 
 spec :: Spec
 spec = describe "Genesis" $ do
