+++ test/Test/Pos/Ssc/GodTossing/Toss/BaseSpec.hs
@@ -10,11 +10,11 @@ import           Control.Lens          (ix)
 import qualified Data.HashMap.Strict   as HM
 import           System.Random         (mkStdGen, randomR)
 
+import           Pos.Arbitrary.Lrc     (GenesisMpcThd, ValidRichmenStake (..))
 import           Pos.Binary            (AsBinary)
 import           Pos.Constants         (genesisBlockVersionData)
 import           Pos.Crypto            (PublicKey, SecretKey, Share,
                                         SignTag (SignCommitment), sign, toPublic)
-import           Pos.Lrc.Arbitrary     (GenesisMpcThd, ValidRichmenStake (..))
 import           Pos.Lrc.Types         (RichmenStake)
 import           Pos.Ssc.GodTossing    (BadCommAndOpening (..), BadCommitment (..),
                                         BadSignedCommitment (..), Commitment,
