+++ test/Test/Pos/Txp/CoreSpec.hs
@@ -18,9 +18,9 @@ import           Test.QuickCheck       (NonNegative (..), Positive (..), Propert
                                         (.&.), (===))
 import           Test.QuickCheck.Gen   (Gen)
 
+import           Pos.Arbitrary.Txp     ()
 import           Pos.Crypto            (hash, whData, withHash)
 import           Pos.Data.Attributes   (mkAttributes)
-import           Pos.Txp.Arbitrary     ()
 import           Pos.Txp.Core          (Tx (..), TxIn (..), TxOut (..), mkTx, topsortTxs)
 import           Pos.Types             (mkCoin)
 import           Pos.Util              (sublistN, _neHead)
