+++ test/Test/Pos/Types/SlottingSpec.hs
@@ -9,7 +9,7 @@ import           Test.Hspec.QuickCheck (prop)
 import           Test.QuickCheck       (Property, (===))
 import           Universum
 
-import           Pos.Core.Arbitrary    ()
+import           Pos.Arbitrary.Core    ()
 import           Pos.Types             (SlotId, flattenSlotId, unflattenSlotId)
 
 spec :: Spec
