+++ txp/Pos/Arbitrary/Txp/Unsafe.hs
@@ -1,10 +1,10 @@
 -- | 'Arbitrary' unsafe instances for some types from Txp types
 
-module Pos.Txp.Arbitrary.Unsafe () where
+module Pos.Arbitrary.Txp.Unsafe () where
 
 import           Universum
 
-import           Pos.Core.Arbitrary.Unsafe ()
+import           Pos.Arbitrary.Core.Unsafe ()
 import           Pos.Txp.Core.Types        (TxOut (..))
 import           Pos.Util.Arbitrary        (ArbitraryUnsafe (..))
